[TASK] Streamline creation of column maps
Currently the datamap factory calls "new ColumnMap" to create
a column map. This is bad as this cannot be mocked at all. A
method that handles the process of creation should be introduced,
also the object manager should be used to create the object.

Important:
This is neither a feature nor a bugfix but as it behaves like a
feature it will not be backported at all!

Releases: 6.1
Resolves: #46354
Change-Id: Ie27da180b05ba11d902da9c11185d41a3a4a8f05
Reviewed-on: https://review.typo3.org/18975
Reviewed-by: Mattias Nilsson
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel