[BUGFIX] Fix PHP warning in BackendUtility::lockRecords
This methods triggers a warning "Illegal String offset"
with PHP 5.4, because $GLOBALS['BE_USER']->user['uid'] is accessed
even if a user is not logged in.

Additionally a delete query is executed in this case which
does not make sense at all.

Properly check if a user is logged in before executing any
functionality in this method.

Fixes: #46361
Releases: 4.5, 4.7, 6.0, 6.1
Change-Id: I370e0a8610a55b3d684bef95d7c6905eaaaacbab
Reviewed-on: https://review.typo3.org/18993
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn