[BUGFIX][EM] Get extension show spinner icon in upper right table
The load mask logic in extension manager stumbles upon the fluid change
of issue #40826 and now always shows a spinner in the "Get extensions"
tab.
This is solved with a small cleanup of the classes and markup added to
this HTML section which anyway are obsolete since #46029.

Change-Id: I0a2e5b26c890ba7ba93bf4abb1f6485084ebf808
Resolves: #46455
Related: #46029
Related: #40826
Releases: 6.1