[BUGFIX][EM] Get extension show spinner icon in upper right table
The load mask logic in extension manager stumbles upon the fluid change
of issue #40826 and now always shows a spinner in the "Get extensions"
tab.
This is solved with a small cleanup of the classes and markup added to
this HTML section which anyway are obsolete since #46029.

Change-Id: I0a2e5b26c890ba7ba93bf4abb1f6485084ebf808
Resolves: #46455
Related: #46029
Related: #40826
Releases: 6.1
Reviewed-on: https://review.typo3.org/19097
Reviewed-by: Jost Baron
Tested-by: Jost Baron
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Alexander Schnitzler
Tested-by: Alexander Schnitzler
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn