[BUGFIX] List module shows thumbs on CEs of type text
The list module renders a thumbnail for content elements as soon
as they have an image attached - independent of the current type
of the content element.

This patch changes the behavior, so that the thumbnail for
records in the list view are only shown, if the thumbnail-
column is visible for the current type.

Change-Id: Id59d7816f91bd37f49f794827826fca1929f7d0a
Resolves: #26753
Releases: 6.1, 6.0