[BUGFIX] List module shows thumbs on CEs of type text
The list module renders a thumbnail for content elements as soon
as they have an image attached - independent of the current type
of the content element.

This patch changes the behavior, so that the thumbnail for
tt_content records in the list view are only shown, if the CType of
the record is either "image" or "textpic". For all other tables, this
does not apply and there will be thumbnails shown, as soon as
there is a thumbnail column defined in TCA which is not empty
for the record to be shown.

Change-Id: Id59d7816f91bd37f49f794827826fca1929f7d0a
Resolves: #26753
Releases: 6.1, 6.0, 4.7, 4.5