[BUGFIX] Exception with EXT:sys_note is installed
If extension sys_note is installed, and the first page does not
have a TypoScript Template with an accordant root flag, a
system exception is thrown, pointing out that the table
"tx_extbase_domain_model_backenduser" could not be found.

The reason for that is:
* the sys_note backend behaviour is based on Extbase
* Extbase is being configured using (front-end) TypoScript
* ext_typoscript_setup.txt files are not processed if there's
  no accordant root template for the current rootline
* the persistence mapping from "be_users" to the mentioned
  "tx_extbase_domain_model_backenduser" would have been defined
  in the external TypoScript file

The solution is to check, whether these external files got
processed during TypoScript parsing and will be processed
explicitely if they have not been integrated yet.

Change-Id: I06a7edecd5cdac9dd7ab3918b281712602ad6b36
Fixes: #46250
Related: #46511
Release: 6.0, 6.1
Reviewed-on: https://review.typo3.org/19134
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader