[BUGFIX] Fix processed files if original has special chars
Filenames of files uploaded in TYPO3 CMS before 6.0.1
can contain problematic characters, because filename
sanitizing of added files was introduced with 6.0.1.

The same happens if files are not uploaded through
the TYPO3 backend and then indexed.

The filenames of processed files are sanitized when
adding them to the storage but the not sanitized
original filename identifier is kept in the database
record of the processed files, which causes wrong
generated paths.

Update the identifier of the processed file along
with all other properties after adding them
to the storage.

Change-Id: I53e4eb42def291ba88ce18209a348b1e2f592185
Resolves: #45221
Related: #42925
Releases: 6.1, 6.0