[BUGFIX] Image rendering of non-existing files throws exception
Rendering an image that has no accordant file in a storage in
the frontend (e.g. in TypoScript) fails with an exception like
"Object with identifier xyz does not exist in storage".

A new ResourceDoesNotExistException is introduced to cover
files and folders and is thrown instead of RuntimeException.
This way, the exception is caught in the ContentObjectRenderer
and the frontend rendering just does not output the error,
which is logged instead.

Change-Id: I0affa15a8f0a50bf30a7176972994f209d18daf4
Fixes: #46535
Releases: 6.0, 6.1