[!!!][BUGFIX] Crop-Scaled images have wrong file content type
The following scenario is given:
* resize a 300dpi TIFF image to PNG using stdWrap/cObj
* the 'fileExtension' configuration is not set
* the processed file csm_... has the file extension PNG
* but the file content type is still TIFF, so the file
  was just renamed, but not converted to PNG

Since the processing task determines the accordant filename
and file extension of the processed file, the image processor
(e.g. ImageMagick) needs to know about that fact as well.

This change modifies the TaskInterface which is rarely to not
at all used in custom extensions implementing individual tasks
to process files.

Change-Id: I089edc01a39663ac91b1b9df168f659c32954fcb
Fixes: #46530
Releases: 6.0, 6.1
Reviewed-on: https://review.typo3.org/19169
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader
Reviewed-on: https://review.typo3.org/19397