[BUG] Preview images don't get updated after overwriting a file
If a file is updated externally, the preview image and the frontend image
are not updated, as the timestamp is not included when creating the
hash.

Additionally, external tools cannot handle special characters when
processing an image which wasn't uploaded by tce_file.php.

That's why the filename creation of the processed file needs to be
cleaned up as well.

Resolves: #44645
Releases: 6.1, 6.0
Change-Id: Iadac7989387ff1de237fe599a582e0e44b4b7540