[BUGFIX] Settings for local drivers are not shown by default
If you create a new storage the LocalDriver is selected by default
(because it is the first registered driver) but the flexform field is not
shown. Because of the hidden configuration field you cannot set up the
base path until you have saved the storage record. Local driver should be
selected by default and configuration field should be displayed every
time.

Change-Id: Ic9c6c1eb7ca3d9f45723dce728558002440ae9da
Fixes: #46555
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/19212
Reviewed-by: Alexander Opitz
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer