[BUGFIX] Segfaults on object comparisons
In the ResourceStorage class, the identity of source and target storage
has to be verified for several operations (copy, move). When the
comparison is done with a simple equality operator (==), the properties
of both objects are compared when the objects are not identical. This
comparison led to a segfault in several environments, presumably because
of a meanwhile fixed PHP bug (https://bugs.php.net/bug.php?id=61557).

To prevent these segfaults on affected PHP versions, the comparison is
extended to check for object identity (===). This will not affect the
behaviour of the code, as there should always be only one object per
storage.

Change-Id: I378c447f75c00859447cf1443f64b81e54161c54
Resolves: #46584
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/19316
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn