[BUGFIX] Upload in backend does not index file
If a file is uploaded within the backend (f.e. file-module)
the file is not automatically indexed. If done within the
file-module the indexing happens when the folder is shown.
If the file is uploaded somehow else using the
ExtendedFileUtility it's not indexed.

Add the indexing to the upload function.

Releases: 6.0, master
Fixes: #46595
Change-Id: I9ea30797390f960a8e382a1f2c456259d3daa60c