[BUGFIX] Default upload folder should be user_upload/
The default folder in FAL is set to temp. However, temp should
be used for temporary files like import/export t3d files,
not uploads by the user.
In 99% of the use case, the uploaded files should be accessible
by the frontend (which temp isn't because of a .htaccess file).
That's why the fileadmin/user_upload/ should always be
created in the install tool, AND the user_upload file
should be the default folder for the local driver.

Resolves: #46604
Releases: 6.1, 6.0
Change-Id: I79319bab4bad76c1af9e001a0a5146fae85ce59f