[FEATURE] Provide flushing cache by table and record tags
As suggested in the wiki it is helpful to tag caches with a combination of
table and record uid to get the cache deleted when the specific record was
edited. But rather that each extension provides its own implementation it
would be better to store this simple function into the core.

Change-Id: I4fb6b49a913d5e428ae2858d0d0a31dc641d2be3
Resolves: #46617
Releases: 6.1