[FEATURE][WIP] common menu options to allow "additionalWhere"
Allow for TMENU/GMENU to have a new option called
"additionalWhere" to narrow down the SQL query put
in the makeMenu function in the PageRepository.

The patch adds a new TypoScript option which is just
sending an extra parameter to the getMenu() function.

This is possible for all HMENU options supported by
the TYPO3 CMS core.

This is a WIP as the unit tests are still missing for that.

Releases: 6.1
Resolves: #46624
Change-Id: Ia96b6aa95049ea6de9ecfd4a5adb69e71ccde2b7