[TASK] Remove unused FAL class ImageProcessingService
In the early stages of the processing functionality, the class
"ImageProcessingService" was added as a proof-of-concept which
failed, as it just moved the cImage functionality and can
only be used in FE (cObj dependency). The new Task interface
for ProcessedFiles serves this purpose since the introduction
of FAL in 6.0.

The file can be removed without any deprecation, as it is
unusable (no API, not used at all).

Releases: 6.1
Resolves: #46625
Change-Id: I8637ca812ed7fd45439cbe85ce3a32ceba946441