[BUGFIX] IRRE children show no field values with useCombination
The IRRE option useCombination enables to use bi-directional
relations with IRRE items. This enables you to re-use IRRE
items multiple times (similar as FAL is working with IRRE).

To prevent the same child record from being used multiple
times, items already used are removed from the select box.
The code removing the item was written to remove an item
from a select box, but (accidentially) also targets a hidden
item which causes the JavaScript to break so that the field
values of the child objects are not filled on expanding.

By adding an additional check, this can be prevented.
For steps to reproduce refer to #46477.

Resolves: #46477
Releases: 6.2
Change-Id: Ie7fd89f6ef887b5d45c734dea8ec6ad9ac2a8d2c
Reviewed-on: https://review.typo3.org/19337
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Zbigniew Jacko
Tested-by: Zbigniew Jacko
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers