[!!!][BUGFIX] Scope of ProcessedFile cannot be modified
A ProcessedFile is used to generate a preview or crop-scaled
version of a File. Actually a Task implementing the
TaskInterface defines the basics and an accordant Helper
object takes care of the modifications to that file.
There are pre-process and post-process signals that seem to
be useless in modifying the scope of a particular task, however.

Imagine a file shall be pre-processed (e.g. color profile
transformation) and that result shall be used for resizing.
The current task implementation is using the original file
directly, however the task actually shall define which is the
source (the original file) the modifications shall be based on.

This change modifies the TaskInterface with a getSourceFile()
method to overcome this limitation.

Change-Id: I280ea9a0bf056f9eee0a06995ab809fdafb8a11a
Fixes: #46795
Releases: 6.0, 6.1