[BUGFIX] Integrity test fails
The integerity test checks if TYPO3_DB is a resource. After switch to
mysqli, this is now an object and thus reflected with the patch now.
At the same time, the $link property - while it changed its type - is
now made protected, and a getter / setter is added.

Change-Id: I5ecdcc27f628352a26a51fc28d5a9d7be5e16315
Resolves: #46805
Related: #36419
Releases: 6.1