[BUGFIX] Update BeUser session expandFolder ref. after folder cmd.
It is not possible to open the file browser when the expandFolder
value of the session moduleData beUser array contains a reference
of an deleted folder.

This patch check if the deleted, renamed or moved folder
is referenced in the array BeUser moduleData and update
the value of expandFolder accordingly.

Change-Id: I53fe988bceaa3404eaf4c7152fa801bcceab8441
Fixes: #46812
Releases: 6.0, 6.1