[BUGFIX] Check & update BeUser moduleData after folder operations
It is not possible to open the ElementBrowser when the expandFolder
value of the session moduleData beUser array contains a reference
of an deleted folder. ResourceDoesNotExistException is thrown.

This patch check if the deleted, renamed or moved folder
is referenced in the BeUser moduleData session storage and update
the value of expandFolder accordingly.

Change-Id: I53fe988bceaa3404eaf4c7152fa801bcceab8441
Fixes: #46812
Releases: 6.0, 6.1