[FEATURE] Allow workspace members to send item back to edit
As all members of a workspace can edit an element in review
they should also be able to send it to editing by changing
the stage. This also allows notification of the original editors
with a stage notification mail.
In the old workspaces this was possible by rejecting an element,
in the current workspaces rejection is not implemented.
So at least this will allow a notification, even though the original
editor has to be tracked by hand.

Change-Id: I0d553570cb9e32bfa49e8821c55b40089ad1bde0
Resolves: #32473
Releases: 6.1
Reviewed-on: https://review.typo3.org/13398
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader