[BUGFIX] Let Upgrade wizard recognize $GLOBALS['TYPO3_CONF_VARS']
The LocalConfiguration wizard rewrites the existing localconf.php to the
new LocalConfiguration.php. To do so, the lines are read and analyzed to
retrieve the value and translate it to the new structure.

Until now, only lines containing the keyword $TYPO3_CONF_VARS are taken
into account. This patch extends the analysis to the keyword
$GLOBALS['TYPO3_CONF_VARS'].

Fixes: #39135
Releases: 6.0, 6.1
Change-Id: I880076d75a312926c8e968ad3ba883c41f364a62
Reviewed-on: https://review.typo3.org/19546
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn