[TASK] Adapt ArrayConstraints in IndexerTest
The tests for a parameter array check for equality of
the parameter array and the according value.
Instead we only wanna know if two keys are present and
set correctly and we do not care about other array values.

This refactors the Test to be more polite and is a follow-up
for #46596.

Change-Id: Ia7b4a0521b1d11b9494de2ff766f11b3c8bd1f60
Releases: 6.0, 6.1
Fixes: #46596