[BUGFIX] Skip some permissions tests if run as root
Some tests in the configuration manager test something is not
writable by manipulating the permissions on the target removing
the w flag. Those tests fail if the suite is run as user root,
this user is always allowed to write. Tests are skipped if run
as root now.

Change-Id: I81d740d7c112df548cf70c76673d433b8f1ff168
Resolves: #46895
Related: #46854
Releases: 6.1