[BUGFIX] Link to not-existing folder results in fatal error
If a folder is removed from the (default) local storage outside of
TYPO3 while still being linked inside a content element, a fatal error
is thrown.

Catch the fatal error and do not render the link.
Rendering the link text, but no link is the default normal behavior of
links in frontent.

Fixes: #48390
Releases: 6.2, 6.1, 6.0
Change-Id: I7ffbf2e055a3cb204dbe61e145ca54ee18b4982b