[BUGFIX] False-Positives in SQL comparison
When one has e.g. a field "foo INT(11) DEFAULT '0' NOT NULL" in
it's ext_tables.sql definition, the comparison will always complain
about that field, since the underlying DB lists this field as "int(11)"
which is lowercased.

This integrates a regex that lowercases the field types before
comparing the field from the definition against the existing field.

Change-Id: If76abbbca56d0ef0ab796a7f4e6bee1197ac39e6
Resolves: #41344
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Reviewed-on: https://review.typo3.org/21353
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters