[WIP][TASK] Set saltedpasswords as required extension
This is the first work to remove challenged and superchallenged
loginSecurityLevel for backend. In new installations rsa and
saltedpasswords are enabled by default. Next thing to do is the
new step for the updateWizard.

In TYPO3 6.0 and former an installation without saltedpasswords was
possible. Now as the system environment check includes tests for OpenSSL
we can force the installation of rsaauth and saltedpasswords as required
extensions.

Change-Id: Id36a48f7463da07197169f63d1862f22a6eca875
Resolves: #39349
Releases: 6.1