[BUGFIX] Disable unsupported mysql modes
If the SQL mode "STRICT_ALL_TABLES" is enabled you
will not be able to save most records. (Enable it in
your my.cnf and try to save a tt_content record)

This patch unsets this option for the mysql session
and triggers a syslog message that this should be
disabled.

Resolves: #20052
Resolves: #18866
Resolves: #18821
Releases: 6.2, 6.1, 6.0, 4.7
Change-Id: Ifd2d7901935f06534a273374bf48266916e23698