[TASK] Remove unused TYPO3 configuration option
The t3InstId option was introduced in version 3.x and never used
by the old EM, and is not even used by the current EM, and also
not with the t3o. As long as there is no functionality
that deals with that option, it should be removed completely again.

Resolves: #46993
Releases: 6.1
Change-Id: Ic506c89375ce012c6ea2bb73649aaa50a2155104