[BUGFIX] EM ConstantsView crashs with dep/faulty types.
The types default and input are deprecated. Removing of this let
the ConstantsView crash, as there is no function configuration for them.

We now readd the config in 6.1 as we have no problems with this lines
in the core and we don't know how much extensions exists which fail.

So removing deprecation in 6.1 message.

Resolves: #46996
Related: #44763
Releases: 6.1, 6.0
Change-Id: Ic9696bb61ac9cbe1618381f42c1c9d8a282cd14c
Reviewed-on: https://review.typo3.org/19698
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn