[TASK] Replace DynTabMenu with jQuery/RequireJS plugin
On the JS code sprint, the task to work on the
DynTabMenu to be abstracted and cleaned, done
via RequireJS + jQuery.

A completely new jQuery plugin is written that
stores the current selected tab ID in the local storage
 (available in IE8+) or falls back to the first tab of the
menu item.

The changes separates logic (JS) from structure
(HTML), and removes plain old JS.

Resolves: #47003
Releases: 6.1
Change-Id: Ib81606cc653ccf7d58105f463bc6a09eb742b7d5