[FEATURE][WIP] Add Drag&Drop Upload to file-list
In TYPO3 6.1 the Flash-Uploader has been removed and the
upload functionality has been reduced to the basic HTML5
multi-upload field.

In the times of HTML5, CSS3 and modern browser our users
expect a way more advanced experience.

This patch adds the possibility to just drag files into the file
list. The dragged files are asynchronously uploaded to the
currently open folder and an reload of the frame is done.

There are still some minor issues and ideas for a even better
user experience, which have to be discussed if they should
be included in the original patch.

Change-Id: I4221f7ff5a06ee11555906f170b37cc08489d875
Task: #47005
Releases: 6.1