[BUGFIX] 47019_Processed folder in a sub folder is shown
In the function fetchFolderListFromDriver of the ResourceStorage class the
processed folder is expected to be in root folder. If it is not the
processed folder is shown in any filelist and gets processed as well. The
function has to split the processed folder identifier and compares it to
the current path.

Change-Id: I7ebde8488938a7000dcfdb99da8c34ee988b666f
Fixes: #47019
Releases: 6.1, 6.0