[BUGFIX] Processed folder in a sub folder is shown
In the function fetchFolderListFromDriver of the ResourceStorage class the
processed folder is expected to be in root folder. If it is not the
case the processing folder is shown in all filelists and gets processed
as well. The function has to split the processed folder identifier and
compare it to the current path. To secure the patch also the missing Unit
Tests were added.

Change-Id: I7ebde8488938a7000dcfdb99da8c34ee988b666f
Fixes: #47019
Releases: 6.2,6.1, 6.0
Reviewed-on: https://review.typo3.org/19732
Reviewed-by: Alexander Opitz
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack
Reviewed-on: https://review.typo3.org/20488