[BUGFIX] ProcessedFile/Thumbnail is always regenerated
A ProcessedFile/Thumbnail is always regenerated even if there
are not modifications to the original file.

The reasons for this spread all over the ProcessedFile handling
and boils down to a new ProcessedFile is being persisted in the
file-system by using a regular ResourceStorage::addFile() call
that results in a plain File object. The indexer then tries to
load this file and will create a new UID for the thumbnail in
sys_file. This uid value is then used to update(!) the
sys_file_processedfile entry - which fails since the uid
was taken from sys_file.

The bug was introduced in
I53e4eb42def291ba88ce18209a348b1e2f592185 were the
processedFile is updated with information of a File object and
the wrong uid is transfered as property.

Fixes: #47140
Releases: 6.1, 6.0
Change-Id: I4252b237d6022485c51ad32a7f956f5afcda8c08