[TASK] Multiple params for user function in typoscript conditions.
Currrently you can use user functions for typoscript conditions and
pass one parameter to your created functions. If you have more than
parameter, you must create a workaround and use an delimiter other
than comma an split it up in your userfunctions to separate them.

But there is no reason to that.

AbstractConditionMatcher line 405 following split up the function
params into an array, but pass only the first param to your function.
Why?

This patch is a fix for that and make it possible to use multiple params
for userfunctions.

Change-Id: Ief2486f6133132303a5925370c1517ee9a98a0b4
Resolves: #47159
Releases: 6.1,6.0