[!!!][BUGFIX] Clear pi_flexform if listType changes
If the list_type field in tt_content changes which means that the
type of plugin changes, the flexform field is cleared.

Reason is that normally no fields are resetted and the fields still
hold the same content. Having switchableControllerActions can lead
to problems when switching between extbase based extensions which
can only be solved by editing the field in the database directly or
by deleting the content element and recreating it.

If 2 plugins would share the same flexform configuration, its
configuration will get lost now when switching the type. The same
happens if a plugin with flexforms switches to e.g. "text" and the
user switches back to the plugin.

Change-Id: I9f2025822dd2d7ec59be66ec9cfbe625a9ad6d4e
Fixes: #31198
Releases: 6.1, 6.0