[BUGFIX] Catch all errors while starting installer
While initializing the installer it can happen that
Exceptions are thrown. As there is no ExceptionHandler
initialized yet, we get a white page and the Fatal error
will be logged.

Mostly it is the problem that typo3conf and/or typo3temp
isn't writeable. But it might something else which throws
an exception.

This patch catches the exception in the initializing
process and shows a page in the layout of the installer.
It shows the exception message, a first hint for the maybe
write permissions problem and a link with the exception number.

Resolves: #11771
Releases: 6.1, 6.0
Change-Id: Ie48ce67f07f350a05c24d6d11bab962ecb11de1c