[BUGFIX] Backend support for native DateTime
The getPlainValue method returns a Unix timestamp for DateTime objects.
Thus it is not possible to store native DateTime objects in
database columns of type Date or DateTime.
This patch passes the columnMap as (optional) argument to
the getPlainValue method and uses the new getDateTimeStorageFormat
method to detect the correct format

Change-Id: I6f8cdd9d131f32e46bf9a99b9c4996d48188312e
Fixes: #47223
Releases: 6.0,6.1
Reviewed-on: https://review.typo3.org/20116
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring