[BUGFIX] Enable readable support for TCA type "group"
The TYPO3 Core supports the mapping of multiple different
datatypes by using an mm-table with a row "tablenames" that
stores the table of the foreign record. As the core only
works with arrays the implementation is fine for it but
Extbase uses objects the data is mapped to.

As there is no definition of which database table is used
for which object a new typoscript configuration has to be
used that defines the mapping. Also a MultipleRelationProxy
has been introduced that enables the creation of relations
to multiple different datatypes.

Example:
/**
 * @var \TYPO3\CMS\Extbase\Persistence\Generic\MultipleReferenceProxy
 */
protectd $items;

The dataMapper recognizes this special type, resolves the
different needed object types and builds an object storage
containing the objects.

Example:
TYPO3\CMS\Extbase\Domain\Model\Category
  items => TYPO3\CMS\Extbase\Persistence\ObjectStorage
    0 => VENDOR\Ext\Domain\Model\Foo
    1 => VENDOR\Ext\Domain\Model\Bar

This patch has to be seen as a bugfix but it is prepared
to be released in TYPO3 6.2. No backports!

Releases: 6.2
Relates: #47362
Relates: #6249
Change-Id: I22c963c59adcccbe56c5995774d6296e6658bc44