[BUGFIX] Select label element from suggest-list
If a sprite is the first child in chosen item from
suggest-list no label is generated.

Therefore select the element depending on the class.
Besides that, e.g. Flexform elements use "invalid" CSS
IDs that contain "[" and "]" characters - using plain
CSS selectors on that does not work and needs to be
escaped first.

Fixes: #47409
Releases: 4.5, 4.7, 6.0, 6.1
Change-Id: I40fad4b6596923cc87889726794f115eb7b58eb5
Reviewed-on: https://review.typo3.org/20071
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader