[TASK] Introduce $GLOBALS['TYPO3_CONF_VARS']['DB']['port']
The [db][host] setting can contain a hostname, followed by a colon and
a port. This was ok with mysql, as it accepted host and port in one
parameter.
With the switch to mysqli, host and port must be given as single
parameters, forcing the core to split host and port. This can fail
especially if the host is an ipv6 address.
In order to separate this better, the new port parameter in
TYPO3_CONF_VARS DB is introduced. If set, the host will not be
splitted by colon anymore.

Resolves: #47455
Releases: 6.1
Change-Id: I3096d166e574d9b6045ac92ed3dc8b2c48ef94d2