[BUGFIX] Warning in backend if a column has no content
The warning is a regression from the ticket #44470. The code
introduced there returns NULL in case a column has no content
elements, but an array is expected by the legacy code in any case.
This leads to a warning in a foreach-loop.

This bugfix adds a type conversion to the loop. This approach was
chosen, because it was already taken in version 6.0. That is also
the reason the bug does not occur there.

Resolves: #47529
Related:  #44470
Releases: 4.7, 4.5

Change-Id: I63b4b26a87f1520139c9d66d0c096a78d39e409c