[BUGFIX] Don't respect sys_language in Backends getObjectByIdentifier
Since the PropertyMapper (on mapping persisted objects) and
the method findByUid of class Repository uses getObjectByIdentifier to
fetch persisted objects and does not create queries on their own any
more the underlying method in Backend must not respect the sysLanguage
to not break the current behaviour.
Adding this setting seems not having any side effects because the method
was only used by the rewritten PropertyMapper.

Fixes: #47585
Fixes: #47586
Relates: #47251
Releases: 6.2, 6.1
Change-Id: I009f8112dd3d96275930132865c0fe5aeb3835f1
Reviewed-on: https://review.typo3.org/20243
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring